Fixed Issue When Adding New Items and Sorting
ngRepeat would erroneously remove items in the view after a sort due to model changes not being detected by angular. This has temporarily been corrected by not adding items to the model until an id has been received from the server. This is not ideal, as we don't want to wait for the server to add an item (might as well do a postback at that point!).